+++ RELEASE_NOTES.md
@@ -3,6 +3,8 @@ __Changes__
 - fixed bugs found in narrative beta testing
 - added error when trying to build ReadsSet mixing PE and SE
 - removed redundant output cells from Set creation Apps
+- made set_provenance() method
+- updated to use KBaseReport.create()
 
 ### Version 1.7.5
 __Changes__